# scrapper/src/main/java/edu/java/controller/dto/request/RemoveLinkRequest.java
package edu.java.controller.dto.request;

import com.fasterxml.jackson.annotation.JsonProperty;
import io.swagger.v3.oas.annotations.media.Schema;
import jakarta.validation.Valid;
import java.net.URI;

public record RemoveLinkRequest(
    @Valid
    @Schema(name = "link", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
    @JsonProperty("link")
    URI uri
) {
}